+++ crates/tauri-cli/src/signer/generate.rs
@@ -3,8 +3,8 @@
 // SPDX-License-Identifier: MIT
 
 use crate::{
-  helpers::updater_signature::{generate_key, save_keypair},
   Result,
+  helpers::updater_signature::{generate_key, save_keypair},
 };
 use clap::Parser;
 use std::path::PathBuf;
@@ -29,7 +29,9 @@ pub struct Options {
 
 pub fn command(mut options: Options) -> Result<()> {
   if options.ci && options.password.is_none() {
-    log::warn!("Generating new private key without password. For security reasons, we recommend setting a password instead.");
+    log::warn!(
+      "Generating new private key without password. For security reasons, we recommend setting a password instead."
+    );
     options.password.replace("".into());
   }
   let keypair = generate_key(options.password).expect("Failed to generate key");
@@ -39,26 +41,37 @@ pub fn command(mut options: Options) -> Result<()> {
       save_keypair(options.force, output_path, &keypair.sk, &keypair.pk)
         .expect("Unable to write keypair");
 
-    println!(
-        "\nYour keypair was generated successfully\nPrivate: {} (Keep it secret!)\nPublic: {}\n---------------------------",
-        display_path(secret_path),
-        display_path(public_path)
-        )
+    println!();
+    println!("Your keypair was generated successfully:");
+    println!("Private: {} (Keep it secret!)", display_path(secret_path));
+    println!("Public: {}", display_path(public_path));
+    println!("---------------------------")
   } else {
-    println!(
-      "\nYour secret key was generated successfully - Keep it secret!\n{}\n\n",
-      keypair.sk
-    );
-    println!(
-          "Your public key was generated successfully:\n{}\n\nAdd the public key in your tauri.conf.json\n---------------------------\n",
-          keypair.pk
-        );
+    println!();
+    println!("Your keys were generated successfully!",);
+    println!();
+    println!("Private: (Keep it secret!)");
+    println!("{}", keypair.sk);
+    println!();
+    println!("Public:");
+    println!("{}", keypair.pk);
   }
 
-  println!("\nEnvironment variables used to sign:");
-  println!("`TAURI_SIGNING_PRIVATE_KEY`  Path or String of your private key");
-  println!("`TAURI_SIGNING_PRIVATE_KEY_PASSWORD`  Your private key password (optional)");
-  println!("\nATTENTION: If you lose your private key OR password, you'll not be able to sign your update package and updates will not work.\n---------------------------\n");
+  println!();
+  println!("Environment variables used to sign:");
+  println!(
+    "- `TAURI_SIGNING_PRIVATE_KEY`: Your private key. For the `build` and `bundle` command it can be either a string or a path to the file, for the `signer sign` command it must be the literal key string"
+  );
+  println!(
+    "- `TAURI_SIGNING_PRIVATE_KEY_PATH`: Path to your private key file, used by the `signer sign` command"
+  );
+  println!(
+    "- `TAURI_SIGNING_PRIVATE_KEY_PASSWORD`:  Your private key password (optional if key has no password)"
+  );
+  println!();
+  println!(
+    "ATTENTION: If you lose your private key OR password, you'll not be able to sign your update package and updates will not work"
+  );
 
   Ok(())
 }