+++ examples/commands/commands.rs
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-License-Identifier: MIT
 
-use tauri::{command, State};
+use tauri::{State, command};
 
 #[command]
 pub fn cmd(_argument: String) {}
@@ -23,5 +23,10 @@ pub fn simple_command(the_argument: String) {
 
 #[command]
 pub fn stateful_command(the_argument: Option<String>, state: State<'_, super::MyState>) {
-  println!("{:?} {:?}", the_argument, state.inner());
+  println!("{:?} {:?}", the_argument, *state);
+}
+
+#[command(rename = "renamed_command_in_mod_new")]
+pub fn renamed_command_in_mod() {
+  println!("renamed command in mod called");
 }