+++ examples/commands/main.rs
@@ -6,13 +6,12 @@
 
 // we move some basic commands to a separate module just to show it works
 mod commands;
-use commands::{cmd, invoke, message, resolver};
+use commands::{cmd, invoke, message, renamed_command_in_mod, resolver};
 
 use serde::Deserialize;
 use tauri::{
-  command,
+  State, Window, command,
   ipc::{Request, Response},
-  State, Window,
 };
 
 #[derive(Debug)]
@@ -51,7 +50,7 @@ async fn async_stateful_command(
   the_argument: Option<String>,
   state: State<'_, MyState>,
 ) -> Result<(), ()> {
-  println!("{:?} {:?}", the_argument, state.inner());
+  println!("{:?} {:?}", the_argument, *state);
   Ok(())
 }
 // ------------------------ Raw future commands ------------------------
@@ -141,7 +140,7 @@ fn stateful_command_with_result(
   the_argument: Option<String>,
   state: State<'_, MyState>,
 ) -> Result<String, MyError> {
-  println!("{:?} {:?}", the_argument, state.inner());
+  println!("{:?} {:?}", the_argument, *state);
   dbg!(the_argument.ok_or(MyError::FooError))
 }
 
@@ -160,7 +159,7 @@ fn stateful_command_with_result_snake(
   the_argument: Option<String>,
   state: State<'_, MyState>,
 ) -> Result<String, MyError> {
-  println!("{:?} {:?}", the_argument, state.inner());
+  println!("{:?} {:?}", the_argument, *state);
   dbg!(the_argument.ok_or(MyError::FooError))
 }
 
@@ -177,7 +176,7 @@ async fn async_stateful_command_with_result(
   the_argument: Option<String>,
   state: State<'_, MyState>,
 ) -> Result<String, MyError> {
-  println!("{:?} {:?}", the_argument, state.inner());
+  println!("{:?} {:?}", the_argument, *state);
   Ok(the_argument.unwrap_or_default())
 }
 
@@ -186,6 +185,11 @@ async fn async_stateful_command_with_result(
 #[command]
 fn command_arguments_wild(_: Window) {
   println!("we saw the wildcard!")
+}
+
+#[command(rename = "renamed_command_new")]
+fn renamed_command() {
+  println!("renamed command called")
 }
 
 #[derive(Deserialize)]
@@ -246,6 +250,8 @@ fn main() {
       future_simple_command,
       async_stateful_command,
       command_arguments_wild,
+      renamed_command,
+      renamed_command_in_mod,
       command_arguments_struct,
       simple_command_with_result,
       async_simple_command_snake,