+++ examples/streaming/main.rs
@@ -109,14 +109,17 @@ fn get_stream_response(
 
       let boundary = random_boundary();
       let boundary_sep = format!("\r\n--{boundary}\r\n");
-      let boundary_closer = format!("\r\n--{boundary}\r\n");
+      let boundary_closer = format!("\r\n--{boundary}--\r\n");
 
-      resp = resp.header(
-        CONTENT_TYPE,
-        format!("multipart/byteranges; boundary={boundary}"),
-      );
+      // `Builder::header` appends, we want to replace the file mime type set earlier
+      if let Some(headers) = resp.headers_mut() {
+        headers.insert(
+          CONTENT_TYPE,
+          HeaderValue::from_str(&format!("multipart/byteranges; boundary={boundary}"))?,
+        );
+      }
 
-      for (end, start) in ranges {
+      for (start, end) in ranges {
         // a new range is being written, write the range boundary
         buf.write_all(boundary_sep.as_bytes())?;
 
@@ -138,6 +141,7 @@ fn get_stream_response(
       // all ranges have been written, write the closing boundary
       buf.write_all(boundary_closer.as_bytes())?;
 
+      resp = resp.status(StatusCode::PARTIAL_CONTENT);
       resp.body(buf)
     }
   } else {